feat(상품): DisplayNamePolicy 추상화
