+++ src/main/java/kitchenpos/products/tobe/domain/policy/DisplayedNamePolicy.java
@@ -1,30 +1,5 @@
 package kitchenpos.products.tobe.domain.policy;
 
-import kitchenpos.products.exception.DisplayedNameException;
-import kitchenpos.products.exception.ProductErrorCode;
-import kitchenpos.products.tobe.domain.DisplayedName;
-import kitchenpos.profanity.ProfanityClient;
-
-public class DisplayedNamePolicy {
-
-    private final ProfanityClient profanityClient;
-
-    public DisplayedNamePolicy(ProfanityClient profanityClient) {
-        this.profanityClient = profanityClient;
-    }
-
-    public void validate(String text) {
-        if (isNullAndEmpty(text)) {
-            throw new DisplayedNameException(ProductErrorCode.NAME_IS_NULL_OR_EMPTY);
-        }
-
-        if (profanityClient.containsProfanity(text)) {
-            throw new DisplayedNameException(ProductErrorCode.NAME_HAS_PROFANITY);
-        }
-    }
-
-    private boolean isNullAndEmpty(String name) {
-        return name == null || name.isBlank();
-    }
-
+public interface DisplayedNamePolicy {
+    void validate(String name);
 }
